+++ commercetools/commercetools-sdk-java-api/src/main/java-predicates-generated/com/commercetools/api/predicates/query/message/OrderPaymentRemovedMessageQueryBuilderDsl.java
@@ -5,98 +5,104 @@ import java.util.function.Function;
 
 import com.commercetools.api.predicates.query.*;
 
-public class OrderPaymentAddedMessageQueryBuilderDsl {
-    public OrderPaymentAddedMessageQueryBuilderDsl() {
+public class OrderPaymentRemovedMessageQueryBuilderDsl {
+    public OrderPaymentRemovedMessageQueryBuilderDsl() {
     }
 
-    public static OrderPaymentAddedMessageQueryBuilderDsl of() {
-        return new OrderPaymentAddedMessageQueryBuilderDsl();
+    public static OrderPaymentRemovedMessageQueryBuilderDsl of() {
+        return new OrderPaymentRemovedMessageQueryBuilderDsl();
     }
 
-    public StringComparisonPredicateBuilder<OrderPaymentAddedMessageQueryBuilderDsl> id() {
+    public StringComparisonPredicateBuilder<OrderPaymentRemovedMessageQueryBuilderDsl> id() {
         return new StringComparisonPredicateBuilder<>(BinaryQueryPredicate.of().left(new ConstantQueryPredicate("id")),
-            p -> new CombinationQueryPredicate<>(p, OrderPaymentAddedMessageQueryBuilderDsl::of));
+            p -> new CombinationQueryPredicate<>(p, OrderPaymentRemovedMessageQueryBuilderDsl::of));
     }
 
-    public LongComparisonPredicateBuilder<OrderPaymentAddedMessageQueryBuilderDsl> version() {
+    public LongComparisonPredicateBuilder<OrderPaymentRemovedMessageQueryBuilderDsl> version() {
         return new LongComparisonPredicateBuilder<>(
             BinaryQueryPredicate.of().left(new ConstantQueryPredicate("version")),
-            p -> new CombinationQueryPredicate<>(p, OrderPaymentAddedMessageQueryBuilderDsl::of));
+            p -> new CombinationQueryPredicate<>(p, OrderPaymentRemovedMessageQueryBuilderDsl::of));
     }
 
-    public DateTimeComparisonPredicateBuilder<OrderPaymentAddedMessageQueryBuilderDsl> createdAt() {
+    public DateTimeComparisonPredicateBuilder<OrderPaymentRemovedMessageQueryBuilderDsl> createdAt() {
         return new DateTimeComparisonPredicateBuilder<>(
             BinaryQueryPredicate.of().left(new ConstantQueryPredicate("createdAt")),
-            p -> new CombinationQueryPredicate<>(p, OrderPaymentAddedMessageQueryBuilderDsl::of));
+            p -> new CombinationQueryPredicate<>(p, OrderPaymentRemovedMessageQueryBuilderDsl::of));
     }
 
-    public DateTimeComparisonPredicateBuilder<OrderPaymentAddedMessageQueryBuilderDsl> lastModifiedAt() {
+    public DateTimeComparisonPredicateBuilder<OrderPaymentRemovedMessageQueryBuilderDsl> lastModifiedAt() {
         return new DateTimeComparisonPredicateBuilder<>(
             BinaryQueryPredicate.of().left(new ConstantQueryPredicate("lastModifiedAt")),
-            p -> new CombinationQueryPredicate<>(p, OrderPaymentAddedMessageQueryBuilderDsl::of));
+            p -> new CombinationQueryPredicate<>(p, OrderPaymentRemovedMessageQueryBuilderDsl::of));
     }
 
-    public CombinationQueryPredicate<OrderPaymentAddedMessageQueryBuilderDsl> lastModifiedBy(
+    public CombinationQueryPredicate<OrderPaymentRemovedMessageQueryBuilderDsl> lastModifiedBy(
             Function<com.commercetools.api.predicates.query.common.LastModifiedByQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.common.LastModifiedByQueryBuilderDsl>> fn) {
         return new CombinationQueryPredicate<>(
             ContainerQueryPredicate.of()
                     .parent(ConstantQueryPredicate.of().constant("lastModifiedBy"))
                     .inner(fn.apply(com.commercetools.api.predicates.query.common.LastModifiedByQueryBuilderDsl.of())),
-            OrderPaymentAddedMessageQueryBuilderDsl::of);
+            OrderPaymentRemovedMessageQueryBuilderDsl::of);
     }
 
-    public CombinationQueryPredicate<OrderPaymentAddedMessageQueryBuilderDsl> createdBy(
+    public CombinationQueryPredicate<OrderPaymentRemovedMessageQueryBuilderDsl> createdBy(
             Function<com.commercetools.api.predicates.query.common.CreatedByQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.common.CreatedByQueryBuilderDsl>> fn) {
         return new CombinationQueryPredicate<>(
             ContainerQueryPredicate.of()
                     .parent(ConstantQueryPredicate.of().constant("createdBy"))
                     .inner(fn.apply(com.commercetools.api.predicates.query.common.CreatedByQueryBuilderDsl.of())),
-            OrderPaymentAddedMessageQueryBuilderDsl::of);
+            OrderPaymentRemovedMessageQueryBuilderDsl::of);
     }
 
-    public LongComparisonPredicateBuilder<OrderPaymentAddedMessageQueryBuilderDsl> sequenceNumber() {
+    public LongComparisonPredicateBuilder<OrderPaymentRemovedMessageQueryBuilderDsl> sequenceNumber() {
         return new LongComparisonPredicateBuilder<>(
             BinaryQueryPredicate.of().left(new ConstantQueryPredicate("sequenceNumber")),
-            p -> new CombinationQueryPredicate<>(p, OrderPaymentAddedMessageQueryBuilderDsl::of));
+            p -> new CombinationQueryPredicate<>(p, OrderPaymentRemovedMessageQueryBuilderDsl::of));
     }
 
-    public CombinationQueryPredicate<OrderPaymentAddedMessageQueryBuilderDsl> resource(
+    public CombinationQueryPredicate<OrderPaymentRemovedMessageQueryBuilderDsl> resource(
             Function<com.commercetools.api.predicates.query.common.ReferenceQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.common.ReferenceQueryBuilderDsl>> fn) {
         return new CombinationQueryPredicate<>(
             ContainerQueryPredicate.of()
                     .parent(ConstantQueryPredicate.of().constant("resource"))
                     .inner(fn.apply(com.commercetools.api.predicates.query.common.ReferenceQueryBuilderDsl.of())),
-            OrderPaymentAddedMessageQueryBuilderDsl::of);
+            OrderPaymentRemovedMessageQueryBuilderDsl::of);
     }
 
-    public LongComparisonPredicateBuilder<OrderPaymentAddedMessageQueryBuilderDsl> resourceVersion() {
+    public LongComparisonPredicateBuilder<OrderPaymentRemovedMessageQueryBuilderDsl> resourceVersion() {
         return new LongComparisonPredicateBuilder<>(
             BinaryQueryPredicate.of().left(new ConstantQueryPredicate("resourceVersion")),
-            p -> new CombinationQueryPredicate<>(p, OrderPaymentAddedMessageQueryBuilderDsl::of));
+            p -> new CombinationQueryPredicate<>(p, OrderPaymentRemovedMessageQueryBuilderDsl::of));
     }
 
-    public StringComparisonPredicateBuilder<OrderPaymentAddedMessageQueryBuilderDsl> type() {
+    public StringComparisonPredicateBuilder<OrderPaymentRemovedMessageQueryBuilderDsl> type() {
         return new StringComparisonPredicateBuilder<>(
             BinaryQueryPredicate.of().left(new ConstantQueryPredicate("type")),
-            p -> new CombinationQueryPredicate<>(p, OrderPaymentAddedMessageQueryBuilderDsl::of));
+            p -> new CombinationQueryPredicate<>(p, OrderPaymentRemovedMessageQueryBuilderDsl::of));
     }
 
-    public CombinationQueryPredicate<OrderPaymentAddedMessageQueryBuilderDsl> resourceUserProvidedIdentifiers(
+    public CombinationQueryPredicate<OrderPaymentRemovedMessageQueryBuilderDsl> resourceUserProvidedIdentifiers(
             Function<com.commercetools.api.predicates.query.message.UserProvidedIdentifiersQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.message.UserProvidedIdentifiersQueryBuilderDsl>> fn) {
         return new CombinationQueryPredicate<>(
             ContainerQueryPredicate.of()
                     .parent(ConstantQueryPredicate.of().constant("resourceUserProvidedIdentifiers"))
                     .inner(fn.apply(
                         com.commercetools.api.predicates.query.message.UserProvidedIdentifiersQueryBuilderDsl.of())),
-            OrderPaymentAddedMessageQueryBuilderDsl::of);
+            OrderPaymentRemovedMessageQueryBuilderDsl::of);
     }
 
-    public CombinationQueryPredicate<OrderPaymentAddedMessageQueryBuilderDsl> paymentRef(
+    public CombinationQueryPredicate<OrderPaymentRemovedMessageQueryBuilderDsl> paymentRef(
             Function<com.commercetools.api.predicates.query.payment.PaymentReferenceQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.payment.PaymentReferenceQueryBuilderDsl>> fn) {
         return new CombinationQueryPredicate<>(ContainerQueryPredicate.of()
                 .parent(ConstantQueryPredicate.of().constant("paymentRef"))
                 .inner(fn.apply(com.commercetools.api.predicates.query.payment.PaymentReferenceQueryBuilderDsl.of())),
-            OrderPaymentAddedMessageQueryBuilderDsl::of);
+            OrderPaymentRemovedMessageQueryBuilderDsl::of);
+    }
+
+    public BooleanComparisonPredicateBuilder<OrderPaymentRemovedMessageQueryBuilderDsl> removedPaymentInfo() {
+        return new BooleanComparisonPredicateBuilder<>(
+            BinaryQueryPredicate.of().left(new ConstantQueryPredicate("removedPaymentInfo")),
+            p -> new CombinationQueryPredicate<>(p, OrderPaymentRemovedMessageQueryBuilderDsl::of));
     }
 
 }